+++ scripts/resources/ItemResource.gd
@@ -1,0 +1,32 @@
+@tool
+extends Resource
+class_name ItemResource
+
+@export_category("Information")
+@export var item_name:String
+@export_enum("Collectable", "ability", "weapon") var item_type = "rescue":
+	set(value):
+		item_type = value
+		notify_property_list_changed()
+@export_category("Statistics")
+## Abilities
+@export var unlock_water := true
+@export var double_jump := true
+## Weapons
+@export var _projectile:ProjectileResource
+@export_category("Sprites")
+@export var pick_up_sprite:CompressedTexture2D
+# Weapons
+@export var weapon_sprite:CompressedTexture2D
+
+
+func _validate_property(property):
+	if property.name in ['unlock_water', 'double_jump'] and item_type != "ability":
+		property.usage = PROPERTY_USAGE_NO_EDITOR
+	if property.name in ['_projectile', 'weapon_sprite'] and item_type != 'weapon':
+		property.usage = PROPERTY_USAGE_NO_EDITOR
+
+"""
+https://github.com/godotengine/godot-proposals/issues/1056
+How to hide properties
+"""